Moved the `y` of the enemydetector lower from the collision detector
I believe the `y`, which is like the + on the area2d, is what is being measured in the enemy stomp detector function.  The body.global_position.y wasn't low enough so the function in the player model that caused the `bounce` was happening before the enemy `stomp detection` so the stomp didn't happen and the bounce did.